Refactor create_structure_geo_hierarchy
When running create_structure_geo_hierarchy we are changing it so that
it processes all structures that can be placed into known jurisdictions
and passses over or skips the ones that cannot.  We also change its
return value to a boolean that will let us know if the structure in question
was processed or not.