Enhance zambia_irs_structures
Following the discussions here https://github.com/onaio/canopy/issues/1502#issuecomment-708360396,
we decided to NOT refresh zambia_irs_structures concurrently.  This means that during its refresh it
will not be able to SELECT from it.

Therefore, to ensure that reporting apps are not affected we are implementing:

- zambia_irs_structures_data: a materialized view that exposes data from zambia_irs_structures which
  will be refreshed concurrently right after zambia_irs_structures is refreshed.
- zambia_irs_structures_geojson: a view (not materialized) that is consumed by reporting apps to expose
  all the data from zambia_irs_structures

This is part of https://github.com/OpenSRP/opensrp-reveal-datawarehouse/issues/33