Fix duplicate event detection in zambia_irs_structures
Change this so that we order by event form end time instead of event date.
This makes dealing with duplicates that occurr on the same day more accurate.

For context, we applied this same fix for namibia_irs_structures and are using
this oppourtinity to do it for Zambia.

Namibia PR for context: https://github.com/OpenSRP/opensrp-reveal-datawarehouse/pull/77