Add plan filters in zambia_irs_structures
The tasks and events selected in zambia_irs_structures did not filter by any plans
and this can be - and has been - a cause of bugs in IRS reports.  This PR fixes that
by ensuring IRS plans are included when calculating zambia_irs_structures.

Additionally, it removes structure information from zambia_irs_structures as this is
not needed - this is instead moved to views that consume zambia_irs_structures.